Readme.md
Windows application/service that will (out of the box) check a list of websites to see if they are up, search the page for a search term, get the space on a local/shared drive and return file/directory counts on a list of directories. All of this is configurable. Plus, more jobs can be added.